+++ composables/useTypedAttrs.ts
@@ -1,0 +1,5 @@
+import { useAttrs } from "#imports";
+
+export default function useTypedAttrs<T extends Record<string, unknown>>(): T {
+  return useAttrs() as T;
+}
